+++ src/routes/provider-show-route/provider-show-route.js
@@ -22,6 +22,7 @@ import {
   packageSortFilterConfig,
   selectionStatusFilterConfig,
   contentTypeFilterConfig,
+  packageAccessFilterConfig,
 } from '../../constants';
 import { SearchSection } from '../../components/search-section';
 import PackageSearchFilters from '../../components/package-search-filters';
@@ -143,6 +144,7 @@ const ProviderShowRoute = ({
           packageSortFilterConfig,
           selectionStatusFilterConfig,
           contentTypeFilterConfig,
+          packageAccessFilterConfig,
         ]}
         {...props}
       />